Address ReadableStream polyfill review feedback
